# Conflicts:
#	presentation/src/main/java/daily/dayo/presentation/view/DropdownMenu.kt

+++ data/src/main/java/daily/dayo/data/datasource/remote/alarm/AlarmResponse.kt
@@ -30,5 +30,5 @@ data class AlarmDto(
     @SerializedName("memberId")
     val memberId: String?,
     @SerializedName("postId")
-    val postId: Int?
+    val postId: Long?
 )
