+++ domain/src/main/java/daily/dayo/domain/usecase/folder/RequestFolderPostListUseCase.kt
@@ -7,6 +7,6 @@ import javax.inject.Inject
 class RequestFolderPostListUseCase @Inject constructor(
     private val folderRepository: FolderRepository
 ) {
-    suspend operator fun invoke(folderId: Int, folderOrder: FolderOrder) =
+    suspend operator fun invoke(folderId: Long, folderOrder: FolderOrder) =
         folderRepository.requestDetailListFolder(folderId, folderOrder)
 }
